fix: fix bugs localstorage
